Fix previous commit changing what Mem[x - C] that had never been written would be on 32bit
Now it is again Mem[x - 10] instead of Mem[x + ffff_fff0], though scarf
still may have both values be created and them compare not equal (as
OperandCtx is not aware of memory space size), while still aliasing the
same data when passed to ExecutionState.

+++ src/exec_state_x86.rs
@@ -793,7 +793,7 @@ impl<'e> State<'e> {
                 return val;
             }
         }
-        ctx.mem_any(size, base, offset as u64)
+        ctx.mem_any(size, base, offset as i32 as i64 as u64)
     }
 
     fn default_mem_value(
